+++ rog-control-center/src/pages/mod.rs
@@ -1,9 +1,11 @@
 mod anime_page;
+mod app_settings;
 mod aura_page;
 mod fan_curve_page;
 mod system_page;
 
 pub use anime_page::*;
+pub use app_settings::*;
 pub use aura_page::*;
 pub use fan_curve_page::*;
 pub use system_page::*;